feat: melhoria nos dados de jogador + remoção da coleção de posição + criação da coleção de estádios + ajuste finos nos campos de algumas coleções

+++ services/dados-historicos/src/main/java/arenaiq/dadoshistoricos/core/dtos/ResultadosCreateDTO.java
@@ -11,7 +11,7 @@ public class ResultadosCreateDTO {
   private Float aproveitamento;
   private Byte golsMarcados;
   private Byte golsSofridos;
-  private Float desempenho;
+  private Byte desempenho;
   private String competicao;
   private String time;
   private String temporada;
