+++ services/dados-historicos/src/main/java/arenaiq/dadoshistoricos/core/dtos/ResultadosDTO.java
@@ -12,7 +12,7 @@ public class ResultadosDTO {
   private Float aproveitamento;
   private Byte golsMarcados;
   private Byte golsSofridos;
-  private Float desempenho;
+  private Byte desempenho;
   private String competicao;
   private String time;
   private String temporada;
